requirements.txt
Django==3.0.8
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-mysql==3.8.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pylint==2.5.3
reportlab==3.4.0
requests==2.18.4
PyMySQL==0.10.0